Added known plaintext attack
In part 1 of the project, the ciphertext that this program has to crack
corresponds to one of the plaintexts in plaintext_dictionary.txt. I
figured that we could start by training the program to recognize one of
these. I've temporarily commented out the section in main.cpp that uses
bigrams to guess. That'll be modified to work in part 2 of the project.